unbundle: do not remove .isolate files, that results in gyp errors
BUG=none
R=maruel@chromium.org, thestig@chromium.org

Review URL: https://codereview.chromium.org/454013002

Cr-Commit-Position: refs/heads/master@{#288682}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288682 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5ebb6ff042c73409792bc2c96899108c794ac4ad

+++ build/linux/unbundle/remove_bundled_libraries.py
@@ -60,6 +60,13 @@ def DoMain(argv):
       if f.endswith('.gyp') or f.endswith('.gypi'):
         continue
 
+      # Deleting .isolate files leads to gyp failures. They are usually
+      # not used by a distro build anyway.
+      # See http://www.chromium.org/developers/testing/isolated-testing
+      # for more info.
+      if f.endswith('.isolate'):
+        continue
+
       if options.do_remove:
         # Delete the file - best way to ensure it's not used during build.
         os.remove(path)
